# backend/src/main/java/io/github/moar0210/assetpulse/alerts/AlertController.java
package io.github.moar0210.assetpulse.alerts;

import io.github.moar0210.assetpulse.identity.AuthenticatedActor;
import io.github.moar0210.assetpulse.security.CorrelationIdFilter;
import jakarta.servlet.http.HttpServletRequest;
import java.util.UUID;
import org.springframework.http.CacheControl;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;

@RestController
@RequestMapping("/api/v1/alerts")
public class AlertController {

    private final AlertQueryService queryService;
    private final AlertCommandService commandService;
    private final AlertStreamService streamService;

    public AlertController(
            AlertQueryService queryService,
            AlertCommandService commandService,
            AlertStreamService streamService) {
        this.queryService = queryService;
        this.commandService = commandService;
        this.streamService = streamService;
    }

    @GetMapping(path = "/stream", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
    public ResponseEntity<SseEmitter> stream(
            @AuthenticationPrincipal AuthenticatedActor actor, HttpServletRequest request) {
        return ResponseEntity.ok()
                .cacheControl(CacheControl.noStore())
                .header("X-Accel-Buffering", "no")
                .body(streamService.subscribe(actor.organisationId(), request.getSession(false)));
    }

    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<AlertListResponse> list(
            @RequestParam(required = false) String limit,
            @AuthenticationPrincipal AuthenticatedActor actor) {
        AlertListRequest request = AlertListRequest.fromQuery(limit);
        return ResponseEntity.ok()
                .cacheControl(CacheControl.noStore())
                .body(queryService.listForOrganisation(actor.organisationId(), request));
    }

    @GetMapping(path = "/{alertId}", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<AlertDetailResponse> detail(
            @PathVariable UUID alertId, @AuthenticationPrincipal AuthenticatedActor actor) {
        return ResponseEntity.ok()
                .cacheControl(CacheControl.noStore())
                .body(queryService.detailForOrganisation(actor.organisationId(), alertId));
    }

    @PostMapping(path = "/{alertId}/acknowledge", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<AlertDetailResponse> acknowledge(
            @PathVariable UUID alertId,
            @AuthenticationPrincipal AuthenticatedActor actor,
            HttpServletRequest request) {
        return ResponseEntity.ok()
                .cacheControl(CacheControl.noStore())
                .body(
                        commandService.acknowledge(
                                actor.organisationId(),
                                actor.userId(),
                                alertId,
                                CorrelationIdFilter.from(request)));
    }

    @PostMapping(path = "/{alertId}/resolve", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<AlertDetailResponse> resolve(
            @PathVariable UUID alertId,
            @AuthenticationPrincipal AuthenticatedActor actor,
            HttpServletRequest request) {
        return ResponseEntity.ok()
                .cacheControl(CacheControl.noStore())
                .body(
                        commandService.resolve(
                                actor.organisationId(),
                                actor.userId(),
                                alertId,
                                CorrelationIdFilter.from(request)));
    }
}