# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AssetPulse Lite API",
    "version": "1.0.0-rc.1",
    "description": "The implemented same-origin API for the AssetPulse Lite demonstration. Organisation, user, and role scope always come from the server-managed session; request input cannot select a tenant or elevate a role.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current same-origin deployment"
    }
  ],
...
  "paths": {
...
    "/api/v1/alerts": {
      "get": {
        "operationId": "listAlerts",
        "summary": "List recent alerts",
        "tags": ["Alerts"],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-assetpulse-roles": ["OPERATIONS_ADMIN", "TECHNICIAN", "VIEWER"],
        "x-assetpulse-tenant-scope": "trusted-session-organisation",
        "parameters": [
          {
            "$ref": "#/components/parameters/ListLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "Bounded tenant-owned alerts.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              },
              "X-Correlation-ID": {
                "$ref": "#/components/headers/CorrelationId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem400"
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          }
        }
      }
    },
    "/api/v1/alerts/stream": {
      "get": {
        "operationId": "streamAlertChanges",
        "summary": "Subscribe to alert invalidation events",
        "description": "Best-effort SSE invalidation only. Logout or session expiry revokes that session's subscriptions. Reconnect requires an active session. Clients refresh authoritative alert state after ready, change, or reconnect events.",
        "tags": ["Alerts"],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-assetpulse-roles": ["OPERATIONS_ADMIN", "TECHNICIAN", "VIEWER"],
        "x-assetpulse-tenant-scope": "trusted-session-organisation",
        "responses": {
          "200": {
            "description": "A ready event followed by zero or more alert-changed events.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              },
              "X-Correlation-ID": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-Accel-Buffering": {
                "description": "Disables reverse-proxy buffering for the event stream.",
                "schema": {
                  "type": "string",
                  "const": "no"
                }
              }
            },
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "UTF-8 Server-Sent Event frames."
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          }
        },
        "x-assetpulse-sse-events": {
          "ready": "#/components/schemas/AlertStreamReadyEvent",
          "alert-changed": "#/components/schemas/AlertStreamChangedEvent"
        }
      }
    },
    "/api/v1/alerts/{alertId}": {
      "get": {
        "operationId": "getAlert",
        "summary": "Read alert detail and history",
        "tags": ["Alerts"],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-assetpulse-roles": ["OPERATIONS_ADMIN", "TECHNICIAN", "VIEWER"],
        "x-assetpulse-tenant-scope": "trusted-session-organisation",
        "parameters": [
          {
            "$ref": "#/components/parameters/AlertId"
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant-owned alert detail.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              },
              "X-Correlation-ID": {
                "$ref": "#/components/headers/CorrelationId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetailResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem400"
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          },
          "404": {
            "$ref": "#/components/responses/Problem404"
          }
        }
      }
    },
    "/api/v1/alerts/{alertId}/acknowledge": {
      "post": {
        "operationId": "acknowledgeAlert",
        "summary": "Acknowledge an open alert",
        "tags": ["Alerts"],
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "x-assetpulse-roles": ["OPERATIONS_ADMIN"],
        "x-assetpulse-tenant-scope": "trusted-session-organisation",
        "parameters": [
          {
            "$ref": "#/components/parameters/AlertId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated alert and immutable history.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              },
              "X-Correlation-ID": {
                "$ref": "#/components/headers/CorrelationId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetailResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem400"
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          },
          "404": {
            "$ref": "#/components/responses/Problem404"
          },
          "409": {
            "$ref": "#/components/responses/Problem409"
          }
        }
      }
    },
    "/api/v1/alerts/{alertId}/resolve": {
      "post": {
        "operationId": "resolveAlert",
        "summary": "Resolve an acknowledged alert",
        "tags": ["Alerts"],
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "x-assetpulse-roles": ["OPERATIONS_ADMIN"],
        "x-assetpulse-tenant-scope": "trusted-session-organisation",
        "parameters": [
          {
            "$ref": "#/components/parameters/AlertId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated alert and immutable history.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              },
              "X-Correlation-ID": {
                "$ref": "#/components/headers/CorrelationId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetailResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem400"
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          },
          "404": {
            "$ref": "#/components/responses/Problem404"
          },
          "409": {
            "$ref": "#/components/responses/Problem409"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ASSETPULSE_SESSION",
        "description": "Server-managed HttpOnly SameSite=Strict session identifier. It carries anonymous CSRF state before login and trusted identity context after login; organisation, actor, and role are loaded from server-owned data."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN",
        "description": "Token returned by GET /api/v1/session/csrf and tied to the current anonymous or authenticated session cookie."
      }
    },
    "headers": {
      "CacheControlNoStore": {
        "description": "Prevents storage of session and tenant-scoped responses.",
        "schema": {
          "type": "string",
          "const": "no-store",
          "example": "no-store"
        }
      },
      "CorrelationId": {
        "description": "Server-generated request correlation identifier. Client-supplied values are ignored.",
        "schema": {
          "$ref": "#/components/schemas/Uuid"
        }
      }
    },
    "parameters": {
...
      "CsrfToken": {
        "name": "X-CSRF-TOKEN",
        "in": "header",
        "required": true,
        "description": "Session-bound mutation token from the CSRF bootstrap route.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "ListLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum results.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
...
      "AlertId": {
        "name": "alertId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "Problem400": {
        "description": "Invalid request.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          },
          "X-Correlation-ID": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem401": {
        "description": "Authentication required or credentials rejected.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          },
          "X-Correlation-ID": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem403": {
        "description": "CSRF verification or role/ownership authorization rejected.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          },
          "X-Correlation-ID": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem404": {
        "description": "Resource absent or outside the trusted organisation/ownership scope.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          },
          "X-Correlation-ID": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem409": {
        "description": "Current state, idempotency key, duplicate, bound, or optimistic version conflict.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          },
          "X-Correlation-ID": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Uuid": {
        "type": "string",
        "format": "uuid"
      },
      "Instant": {
        "type": "string",
        "format": "date-time"
      },
...
      "AlertStatus": {
        "type": "string",
        "enum": ["OPEN", "ACKNOWLEDGED", "RESOLVED"]
      },
...
      "MeasurementType": {
        "type": "string",
        "enum": ["TEMPERATURE"]
      },
      "MeasurementUnit": {
        "type": "string",
        "enum": ["CELSIUS"]
      },
      "ThresholdComparison": {
        "type": "string",
        "enum": ["GREATER_THAN_OR_EQUAL_TO"]
      },
...
      "Problem": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "correlationId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*$"
          },
          "correlationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "AlertAssetContext": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "assetCode", "name"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "assetCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AlertSensorContext": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "sensorKey", "name", "measurementType", "unit"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "sensorKey": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "measurementType": {
            "$ref": "#/components/schemas/MeasurementType"
          },
          "unit": {
            "$ref": "#/components/schemas/MeasurementUnit"
          }
        }
      },
      "AlertRuleContext": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "ruleCode",
          "name",
          "comparison",
          "thresholdValue",
          "cooldownSeconds"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "ruleCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "comparison": {
            "$ref": "#/components/schemas/ThresholdComparison"
          },
          "thresholdValue": {
            "type": "number"
          },
          "cooldownSeconds": {
            "type": "integer",
            "minimum": 0,
            "maximum": 604800
          }
        }
      },
      "AlertContextResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["asset", "sensor", "thresholdRule"],
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AlertAssetContext"
          },
          "sensor": {
            "$ref": "#/components/schemas/AlertSensorContext"
          },
          "thresholdRule": {
            "$ref": "#/components/schemas/AlertRuleContext"
          }
        }
      },
      "AlertSummaryResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "status",
          "occurrenceCount",
          "lastOccurredAt",
          "context"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "status": {
            "$ref": "#/components/schemas/AlertStatus"
          },
          "occurrenceCount": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "lastOccurredAt": {
            "$ref": "#/components/schemas/Instant"
          },
          "context": {
            "$ref": "#/components/schemas/AlertContextResponse"
          }
        }
      },
      "AlertListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["alerts", "limit"],
        "properties": {
          "alerts": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/AlertSummaryResponse"
            }
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        }
      },
      "ActorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "displayName"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
      "AlertHistoryResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "sequenceNumber",
          "fromStatus",
          "toStatus",
          "actor",
          "transitionedAt"
        ],
        "properties": {
          "sequenceNumber": {
            "type": "integer",
            "minimum": 1
          },
          "fromStatus": {
            "$ref": "#/components/schemas/AlertStatus"
          },
          "toStatus": {
            "$ref": "#/components/schemas/AlertStatus"
          },
          "actor": {
            "$ref": "#/components/schemas/ActorResponse"
          },
          "transitionedAt": {
            "$ref": "#/components/schemas/Instant"
          }
        }
      },
      "AlertDetailResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "status",
          "occurrenceCount",
          "firstOccurredAt",
          "lastOccurredAt",
          "cooldownUntil",
          "createdAt",
          "updatedAt",
          "context",
          "history"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "status": {
            "$ref": "#/components/schemas/AlertStatus"
          },
          "occurrenceCount": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "firstOccurredAt": {
            "$ref": "#/components/schemas/Instant"
          },
          "lastOccurredAt": {
            "$ref": "#/components/schemas/Instant"
          },
          "cooldownUntil": {
            "$ref": "#/components/schemas/Instant"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Instant"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Instant"
          },
          "context": {
            "$ref": "#/components/schemas/AlertContextResponse"
          },
          "history": {
            "type": "array",
            "maxItems": 2,
            "items": {
              "$ref": "#/components/schemas/AlertHistoryResponse"
            }
          }
        }
      },
      "AlertChangeEvent": {
        "type": "object",
        "additionalProperties": false,
        "required": ["alertId", "changeType"],
        "properties": {
          "alertId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "changeType": {
            "type": "string",
            "enum": ["OCCURRENCE_RECORDED", "STATUS_CHANGED"]
          }
        }
      },
      "AlertStreamReadyEvent": {
        "type": "object",
        "additionalProperties": false,
        "required": ["event", "data"],
        "properties": {
          "event": {
            "const": "ready"
          },
          "data": {
            "type": "object",
            "additionalProperties": false,
            "maxProperties": 0
          }
        }
      },
      "AlertStreamChangedEvent": {
        "type": "object",
        "additionalProperties": false,
        "required": ["event", "data"],
        "properties": {
          "event": {
            "const": "alert-changed"
          },
          "data": {
            "$ref": "#/components/schemas/AlertChangeEvent"
          }
        }
      },
...
    }
  }
}